Introduce `OracleEnhanced::Lob` not to modify `ActiveRecord::Base` directly
To load LOB data from fixtures, Oracle enhanced adapter
has modified `ActiveRecore::Base`, which can be avoided
by introducing `OracleEnhanced::Lob` module using `ActiveSupport::Concern`
